% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annulus.r
\name{annulus}
\alias{annulus}
\title{Plot an annulus on an existing plot.}
\usage{
annulus(x, y, inner, outer, deg = c(0, 360), n = 1000, col = "black",
  density = NULL, angle = 45, border = col, force0 = FALSE, ...)
}
\arguments{
\item{x}{Numeric, x-coordinate of center of annulus.}

\item{y}{Numeric, y-coordinate of center of annulus.}

\item{inner}{Numeric, inner radius.}

\item{outer}{Numeric, outer radius.}

\item{deg}{Two-element numeric list, degrees across which to draw the annulus (0 is "north" and increasing values go clockwise). Default is \code{c(0, 360)}.}

\item{n}{Integer, number of vertices used to approximate a circle.}

\item{col}{Character or integer, color used to fill annulus.  Note that if \code{col} and \code{border} are different colors and \code{deg} is \code{c(0, 360)} (the default), then you will able to see the line connecting the inner and outer borders.}

\item{density}{Numeric, the density of shading lines, in lines per inch. The default value of \code{NULL} causes no shading lines to be drawn. A zero value of density means no shading nor filling whereas negative values and \code{NA} suppress shading (and so allow color filling).}

\item{angle}{Numeric, the slope of shading lines, given as an angle in degrees (counter-clockwise).}

\item{border}{Character or integer, color used to draw annulus border.  Note that if \code{col} and \code{border} are different colors and \code{deg} is \code{c(0, 360)} (the default), then you will able to see the line connecting the inner and outer borders.}

\item{force0}{Logical, if \code{TRUE} then negative values of inner and outer are coerced to equal 0. If \code{FALSE} then throws an error.}

\item{...}{Arguments to send to \code{\link[graphics]{polygon}}.}
}
\value{
Nothing (side-effect is to add an annulus to an existing plot).
}
\description{
This function adds an annulus to an existing plot.
}
\examples{
par(pty='s')
plot(0, 0, xlim=c(-1, 1), ylim=c(-1, 1))
annulus(x=0.3, y=-0.2, inner=0.3, outer=1, col='cornflowerblue')
annulus(x=-0.3, y=0.2, inner=0.2, outer=0.4, deg=c(0, 135), border='red')
}
\seealso{
\code{\link[legendary]{annulusSeg}}, \code{\link[graphics]{polygon}}
}
