% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annulusSeg.r
\name{annulusSeg}
\alias{annulusSeg}
\title{Plot an annulus with different segments on an existing plot}
\usage{
annulusSeg(x, y, inner, outer, deg = NULL, n = 1000, col = "black",
  border = col, force0 = FALSE, ...)
}
\arguments{
\item{x}{Numeric, x-coordinate of center of annulus.}

\item{y}{Numeric, y-coordinate of center of annulus.}

\item{inner}{Numeric vector, inner radius of each segment.}

\item{outer}{Numeric vector, outer radius of each segment.}

\item{n}{Integer, number of vertices used to approximate a circle.}

\item{col}{Character or integer vector, colors used to fill segments.}

\item{border}{Character or integer vector, color used to draw segment borders.}

\item{force0}{Logical, if \code{TRUE} then negative values of inner and outer are coerced to equal 0. If \code{FALSE} (default) then throws an error.}

\item{...}{Arguments to send to \code{\link[graphics]{polygon}}.}
}
\value{
Nothing (side-effect is to add a segmented annulus to an existing plot).
}
\description{
This function adds an annulus to an existing plot. It is very similar to the \code{\link[legendary]{annulus}} function except that the annulus can be composed of one or more "segments" with different formatting (color, etc.).
}
\examples{
inner <- c(0.2, 0.4, 0.7)
outer <- c(0.5, 0.6, 0.71)
col <- c('red', 'blue', 'green')
par(pty='s')
plot(0, 0, xlim=c(-1, 1), ylim=c(-1, 1), col='white')
annulusSeg(x=0, y=0, inner, outer, col=col, border='black')
}
\seealso{
\code{\link[legendary]{annulus}}, \code{\link[graphics]{polygon}}
}
