% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.r
\name{arc}
\alias{arc}
\title{Plot an arc on an existing plot}
\usage{
arc(x, y, radius, deg, n = 1000L, ...)
}
\arguments{
\item{x}{Numeric, x-coordinate of center of arc.}

\item{y}{Numeric, y-coordinate of center of arc.}

\item{radius}{Numeric, radius of arc.}

\item{deg}{Two-element numeric vector, degrees across which to draw arc (0 is up, 90 right, 180 down, and 270 left).}

\item{...}{Arguments to send to \code{\link[graphics]{lines}}.}
}
\value{
Nothing (side-effect is to add an arc to an existing plot).
}
\description{
This function adds an arc to an existing plot.
}
\examples{
par(pty='s')
plot(0, 0, xlim=c(-1, 1), ylim=c(-1, 1))
arc(x=0.3, y=-0.2, radius=0.6, deg=c(0, 135), col='cornflowerblue')
arc(x=0, y=0, radius=0.3, deg=c(90, 360), col='red', lty='dotted')
arc(x=1, y=0, radius=2, deg=c(225, 325), col='green', lwd=2)
}
\seealso{
\code{\link[graphics]{lines}}
}
