% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_locateElement.r
\name{.locateElement}
\alias{.locateElement}
\title{Locate a plot element (e.g., a legend)}
\usage{
.locateElement(x, y = NULL, inset = 0, width = 0.2, height = 0.5)
}
\arguments{
\item{x}{Numeric or character. Describes the location of the legend. This is a numeric value (in which case \code{y} must also be supplied) indicating the x-coordinate of the top left of the box surrounding the legend. Alternatively, it is a character describing the position of the box surrounding the legend relative to the existing plot (\code{'topleft'}, \code{'topright'}, \code{'bottomleft'}, \code{'bottomright'}, \code{'top'}, \code{'bottom'}, \code{'left'}, \code{'right'}, or \code{'center'}).}

\item{y}{Numeric or \code{NULL}.}

\item{inset}{Numeric. If \code{x} is a word describing the position of the legend, then this is the degree to which the legend is inset (or outset, if negative) relative to the figure's border. If two values are supplied then the first pertains to the horizontal offset and the second the vertical offset.}

\item{width}{Numeric. Scaling factor for box width.}

\item{height}{Numeric. Scaling factor for box height.}

\item{return}{Coordinates of the upper left of the element.}
}
\description{
This is a private function that locates a plot element.
}
\keyword{internal}
