% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyCorner.r
\name{emptyCorner}
\alias{emptyCorner}
\title{Find the corner of a plot region that is most empty of plot elements}
\usage{
emptyCorner(x, by = "distance", onlyIn = TRUE, least = TRUE)
}
\arguments{
\item{x}{A two-column matrix or data frame with points representing plot items.}

\item{by}{Character, method by which to identify the "most empty" quadrat. Any of (partial matching is OK, and case is ignored):
\itemize{
        \item \code{'distance'} (default): Find the corner that has the greatest distance between it and the closest point.
        \item \code{'density'}: Find the corner that has the least density of points.
        \item \code{'dd'}: Find the corner that has the least density of closest points.
}}

\item{onlyIn}{Logical, if \code{TRUE} (default), only consider plot items that are in the plot region.}

\item{least}{Logical, if \code{TRUE} (default), then indicate the least-occupied corner. If \code{FALSE} then indicate the most-occupied corner.}
}
\value{
Integer: \code{1}: top left; \code{2}: top right; \code{3}: bottom left; \code{4}: bottom right.
}
\description{
This function identifies the corner of a plot region that is the least occupied by plot elements.
}
\details{
Note that you must already have a plot made for this to work.
}
\examples{
# generate some numbers
set.seed(123)
x <- c(4, rnorm(100), runif(200, -2, 0))
y <- c(2, rnorm(100), runif(200, -2, 0))
xlim <- c(min(x), 1.5 * max(x))
plot(x, y, xlim=xlim)
xy <- cbind(x, y)
emptyCorner(xy)
emptyCorner(xy, by='density')
emptyCorner(xy, by='dd')
}
