% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFig.r
\name{getFig}
\alias{getFig}
\title{Return coordinates for par(fig=xxxx)}
\usage{
getFig(layout, n)
}
\arguments{
\item{layout}{Matrix representing the positioning of subplots. See \code{\link[graphics]{layout}}.}

\item{n}{Integer, indicating which suboplot to plot (corresponds to value(s) in \code{layout}).}
}
\value{
A four-element numeric vector.
}
\description{
This function can be used with the \code{\link[graphics]{layout}} function for defining the positioning of multiple plots on a single plotting device. Given a matrix used in the \code{layout} function and the number indicating which a particular subplot area, it returns the min/max x- and y-coordinates for the subplot relative to the entire plottig region.
}
\examples{
lo <- matrix(c(1, 2, 3), nrow=1)
layout(lo)
subplot1 <- getFig(lo, 1)
subplot2 <- getFig(lo, 2)
subplot3 <- getFig(lo, 3)
par(fig=subplot2, new=FALSE)
plot(1, 1, main='Subplot 2')
par(fig=subplot1, new=TRUE)
plot(1, 1, main='Subplot 1')
par(fig=subplot3, new=TRUE)
plot(1, 1, main='Subplot 3')
}
