% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelFig.r
\name{labelFig}
\alias{labelFig}
\title{Adds a label to the upper left corner of a plot}
\usage{
labelFig(label, adj = 0.05, ...)
}
\arguments{
\item{label}{Character.}

\item{adj}{A single number or two numbers. If a single number, then this is an adjustment factor that affects the position of the label in the y-direction relative to the default position. Positive connotes higher, negative lower. If this is two values, then the first adjusts in the x-direction (negative connotes left, positive right), and the second adjust in the y-direction.}

\item{...}{Other arguments to send to \code{\link[graphics]{text}}.}
}
\value{
Nothing (adds text to a plot).
}
\description{
Adds a label to the upper left corner of a plot.
}
\examples{
par(mfrow=c(1, 2))
plot(1, 1)
labelFig('A: One')
plot(2, 2)
labelFig('B: Two')
}
