% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancakes.r
\name{pancakes}
\alias{pancakes}
\title{Plots two horizontal barplots back-to-back.}
\usage{
pancakes(x1, x2, sameScale = TRUE, xlim1 = NULL, xlim2 = NULL,
  tickLabelOffset = -0.05, xlab1 = "x1", xlab2 = "x2", xlab = NULL,
  xlabOffset = -0.12, tickSize = 0.01, labels = NULL,
  labelSide = "both", cexLab = 1, col1 = "gray70", col2 = "gray30",
  border1 = "black", border2 = "black", ...)
}
\arguments{
\item{x1}{Numeric, vector of values for the left-hand side of the plot. Values are plotted such that the first is represented by the lowest bar, the second the next-higher bar, and so on.}

\item{x2}{Numeric, vector of values for the right-hand side of the plot, must be same length as \code{x1}.}

\item{sameScale}{Logical, if \code{TRUE} (default) then use the same scale for both sides of the plot. If \code{FALSE}, then each side will have its own scale determined automatically \emph{unless} \code{xlim1} and/or \code{xlim2} are non-\code{NULL}.}

\item{xlim1, xlim2}{Two numeric values defining the x-axis limits for \code{x1} and \code{x2}. Alternatively, if one or both are \code{NULL} (default), then the scale is determined automatically for either or both.}

\item{tickLabelOffset}{Numeric, offset (in y-direction) of labels for x-axis tick marks.}

\item{xlab1, xlab2}{Characters, names of each x-axis label. Note that by default these are used but if \code{xlab} is non-\code{NULL}, then it is used insetad.}

\item{xlab}{Character, name of x-axis label.}

\item{xlabOffset}{Numeric, offset (in y-direction) of x-axis label(s).}

\item{tickSize}{Numeric, relative size of tick marks. Positive numbers grow them longer from the bottom of the x-axis, negative upward.}

\item{labels}{Vector of same length as \code{x1} or \code{x2} with labels for bars. Can be numeric, character, integers, or anything else that prints. Default is \code{NULL}, in which case no labels are added.}

\item{labelSide}{Character, either \code{left}, \code{right}, or \code{both}, indicating side on which bar labels are drawn.}

\item{cexLab}{Either \code{NA} (default) or a numeric > 0. Size of label text.}

\item{col1, col2}{Colors of bars.}

\item{border1, border2}{Border colors of bars.}

\item{...}{Arguments to pass to \code{\link[graphics]{par}} or \code{\link[graphics]{plot}}.}

\item{labelOffset}{Numeric, offset (from border or plot region) for positioning of labels. Positive numbers move labels outward, negative inward.}
}
\value{
None. Side-effect is to generate a plot.
}
\description{
This function plots two horizontal barplots back-to-back.
}
\details{
If you want to add elements (e.g., arrows) to the plot, then note that the x-axis is scaled fromm -1 to 1 with the two sets of bar plots abutting one another at x = 0. The y-axis is scaled from 0 to 1.
}
\examples{
x1 <- 10:1
x2 <- sample(11:20, 10)
labels <- LETTERS[10:1]
pancakes(x1, x2, labels = labels)
pancakes(x1, x2, sameScale = FALSE, labels=labels)
}
\seealso{
\code{\link{barplot}}
}
