% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welfare.r
\docType{data}
\name{welfare}
\alias{welfare}
\title{GDP, population, and HIV prevalence for the countries of the world}
\format{
A data frame with GDP, population, and HIV prevalence for the countries of the world.
}
\source{
\href{http://datatopics.worldbank.org/world-development-indicators/}{World Bank Development Index}
}
\usage{
data(welfare)
}
\description{
GDP, population, and HIV prevalence for the countries of the world
}
\examples{
data(welfare)
}
\keyword{datasets}
